[695] fix: add darkFactoryOrchestrator to ipc test mocks
